refactor: preprocessing is a different step
